@@ -1,1 +1,1 @@
-__version__ = "0.41.2"
+__version__ = "0.45.2"
