@@ -21,21 +21,21 @@ def test_https_redirect_middleware(test_client_factory: TestClientFactory) -> No
     assert response.status_code == 200
 
     client = test_client_factory(app)
-    response = client.get("/", allow_redirects=False)
+    response = client.get("/", follow_redirects=False)
     assert response.status_code == 307
     assert response.headers["location"] == "https://testserver/"
 
     client = test_client_factory(app, base_url="http://testserver:80")
-    response = client.get("/", allow_redirects=False)
+    response = client.get("/", follow_redirects=False)
     assert response.status_code == 307
     assert response.headers["location"] == "https://testserver/"
 
     client = test_client_factory(app, base_url="http://testserver:443")
-    response = client.get("/", allow_redirects=False)
+    response = client.get("/", follow_redirects=False)
     assert response.status_code == 307
     assert response.headers["location"] == "https://testserver/"
 
     client = test_client_factory(app, base_url="http://testserver:123")
-    response = client.get("/", allow_redirects=False)
+    response = client.get("/", follow_redirects=False)
     assert response.status_code == 307
     assert response.headers["location"] == "https://testserver:123/"
