@@ -2,9 +2,10 @@ from __future__ import annotations
 
 import datetime as dt
 import time
+from collections.abc import AsyncGenerator, AsyncIterator, Iterator
 from http.cookies import SimpleCookie
 from pathlib import Path
-from typing import Any, AsyncIterator, Iterator
+from typing import Any
 
 import anyio
 import pytest
@@ -12,7 +13,7 @@ import pytest
 from starlette import status
 from starlette.background import BackgroundTask
 from starlette.datastructures import Headers
-from starlette.requests import Request
+from starlette.requests import ClientDisconnect, Request
 from starlette.responses import FileResponse, JSONResponse, RedirectResponse, Response, StreamingResponse
 from starlette.testclient import TestClient
 from starlette.types import Message, Receive, Scope, Send
@@ -89,7 +90,7 @@ def test_redirect_response_content_length_header(
         await response(scope, receive, send)
 
     client: TestClient = test_client_factory(app)
-    response = client.request("GET", "/redirect", allow_redirects=False)
+    response = client.request("GET", "/redirect", follow_redirects=False)
     assert response.url == "http://testserver/redirect"
     assert response.headers["content-length"] == "0"
 
@@ -265,7 +266,7 @@ async def test_file_response_on_head_method(tmp_path: Path) -> None:
             assert "content-disposition" in headers
             assert "last-modified" in headers
             assert "etag" in headers
-        elif message["type"] == "http.response.body":
+        elif message["type"] == "http.response.body":  # pragma: no branch
             assert message["body"] == b""
             assert message["more_body"] is False
 
@@ -336,6 +337,22 @@ def test_file_response_with_method_warns(tmp_path: Path) -> None:
         FileResponse(path=tmp_path, filename="example.png", method="GET")
 
 
+def test_file_response_with_range_header(tmp_path: Path, test_client_factory: TestClientFactory) -> None:
+    content = b"file content"
+    filename = "hello.txt"
+    path = tmp_path / filename
+    path.write_bytes(content)
+    etag = '"a_non_autogenerated_etag"'
+    app = FileResponse(path=path, filename=filename, headers={"etag": etag})
+    client = test_client_factory(app)
+    response = client.get("/", headers={"range": "bytes=0-4", "if-range": etag})
+    assert response.status_code == status.HTTP_206_PARTIAL_CONTENT
+    assert response.content == content[:5]
+    assert response.headers["etag"] == etag
+    assert response.headers["content-length"] == "5"
+    assert response.headers["content-range"] == f"bytes 0-4/{len(content)}"
+
+
 def test_set_cookie(test_client_factory: TestClientFactory, monkeypatch: pytest.MonkeyPatch) -> None:
     # Mock time used as a reference for `Expires` by stdlib `SimpleCookie`.
     mocked_now = dt.datetime(2037, 1, 22, 12, 0, 0, tzinfo=dt.timezone.utc)
@@ -375,6 +392,18 @@ def test_set_cookie_path_none(test_client_factory: TestClientFactory) -> None:
     response = client.get("/")
     assert response.text == "Hello, world!"
     assert response.headers["set-cookie"] == "mycookie=myvalue; SameSite=lax"
+
+
+def test_set_cookie_samesite_none(test_client_factory: TestClientFactory) -> None:
+    async def app(scope: Scope, receive: Receive, send: Send) -> None:
+        response = Response("Hello, world!", media_type="text/plain")
+        response.set_cookie("mycookie", "myvalue", samesite=None)
+        await response(scope, receive, send)
+
+    client = test_client_factory(app)
+    response = client.get("/")
+    assert response.text == "Hello, world!"
+    assert response.headers["set-cookie"] == "mycookie=myvalue; Path=/"
 
 
 @pytest.mark.parametrize(
@@ -542,6 +571,39 @@ async def test_streaming_response_stops_if_receiving_http_disconnect() -> None:
     assert not cancel_scope.cancel_called, "Content streaming should stop itself."
 
 
+@pytest.mark.anyio
+async def test_streaming_response_on_client_disconnects() -> None:
+    chunks = bytearray()
+    streamed = False
+
+    async def receive_disconnect() -> Message:
+        raise NotImplementedError
+
+    async def send(message: Message) -> None:
+        nonlocal streamed
+        if message["type"] == "http.response.body":
+            if not streamed:
+                chunks.extend(message.get("body", b""))
+                streamed = True
+            else:
+                raise OSError
+
+    async def stream_indefinitely() -> AsyncGenerator[bytes, None]:
+        while True:
+            await anyio.sleep(0)
+            yield b"chunk"
+
+    stream = stream_indefinitely()
+    response = StreamingResponse(content=stream)
+
+    with anyio.move_on_after(1) as cancel_scope:
+        with pytest.raises(ClientDisconnect):
+            await response({"asgi": {"spec_version": "2.4"}}, receive_disconnect, send)
+    assert not cancel_scope.cancel_called, "Content streaming should stop itself."
+    assert chunks == b"chunk"
+    await stream.aclose()
+
+
 README = """\
 # BáiZé
 
@@ -700,7 +762,7 @@ async def test_file_response_multi_small_chunk_size(readme_file: Path) -> None:
     async def send(message: Message) -> None:
         if message["type"] == "http.response.start":
             start_message.update(message)
-        elif message["type"] == "http.response.body":
+        elif message["type"] == "http.response.body":  # pragma: no branch
             received_chunks.append(message["body"])
 
     await app({"type": "http", "method": "get", "headers": [(b"range", b"bytes=0-15,20-35,35-50")]}, receive, send)
